added validation for categories, commentaries, guides and more

+++ serverApp/src/main/java/paper/services/CategoryService.java
@@ -1,7 +1,9 @@
 package paper.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import paper.entities.Category;
 import paper.repository.CategoryRepository;
 
@@ -17,13 +19,24 @@ public class CategoryService {
         this.categoryRepository = categoryRepository;
     }
 
-    // TODO: validation
     public void createCategory(Category category) {
+        if (category == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The request body is null");
+        }
+
         categoryRepository.save(category);
     }
 
     public void deleteCategory(Category category) {
-        categoryRepository.deleteById(category.getName());
+        if (category == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The request body is null");
+        }
+
+        if (categoryRepository.existsById(category.getName())) {
+            categoryRepository.deleteById(category.getName());
+        } else {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The category by given name does not exist");
+        }
     }
 
     public List<Category> getCategories() {
